% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxonomic_resources.R
\name{clear_cached_resources}
\alias{clear_cached_resources}
\title{Clear cached taxonomic resources}
\usage{
clear_cached_resources()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Removes any taxonomic resources that have been cached in memory during the
current R session. After calling this function, the next call to
\code{\link[=load_taxonomic_resources]{load_taxonomic_resources()}} will re-download and re-process the data.
}
\details{
This is useful if you want to force a reload of the resources, for example
after updating the package or switching to a different version.
}
\examples{
clear_cached_resources()
}
