% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_diversity.R
\name{create_species_state_origin_matrix}
\alias{create_species_state_origin_matrix}
\title{Process Geographic Data and Return State Level Species Origin and Diversity Counts}
\usage{
create_species_state_origin_matrix(
  type_of_data = "stable",
  version = default_version()
)
}
\arguments{
\item{type_of_data}{A character string indicating the type of data to be used. The default is "stable".  "current" downloads the file from APC directly and remakes the calculations from the most recent version.}

\item{version}{Version of the database (for the stable option only)}
}
\value{
A data frame with columns representing each state and rows representing each species. The values in each cell represent the origin of the species in that state.
}
\description{
This function processes the geographic data available in the current Australian Plant Census and returns state level diversity for native, introduced and more complicated species origins.
}
\examples{
create_species_state_origin_matrix()

}
