% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{create_taxonomic_update_lookup}
\alias{create_taxonomic_update_lookup}
\title{Create a lookup table to help fix the taxonomy for a list of Australian plant species}
\usage{
create_taxonomic_update_lookup(
  species_list,
  fuzzy_matching = FALSE,
  version_number = default_version(),
  full = FALSE
)
}
\arguments{
\item{species_list}{A list of Australian plant species that needs to be reconciled with current taxonomy.}

\item{fuzzy_matching}{A logical value indicating whether fuzzy matching should be used to align the species names. Default is \code{FALSE}.}

\item{version_number}{The version number of the dataset to use. Default is \code{"0.0.1.9000"}.}

\item{full}{logical for whether the full lookup table is returned or just the two key columns}
}
\value{
A lookup table containing the original species names, the aligned species names, and additional taxonomic information such as taxon IDs and genera.
}
\description{
This function takes a list of Australian plant species that needs to be reconciled with current taxonomy and generates a lookup table to help fix the taxonomy. The lookup table contains the original species names, the aligned species names, and additional taxonomic information such as taxon IDs and genera.
}
\examples{
create_taxonomic_update_lookup(c("Eucalyptus regnans", "Acacia melanoxylon", 
"Banksia integrifolia","Not a species"))

}
