% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access.R
\name{dataset_del}
\alias{dataset_del}
\title{Delete a dataset release from GitHub}
\usage{
dataset_del(version, path = NULL)
}
\arguments{
\item{version}{The version number of the dataset release to delete.}

\item{path}{(optional) The path to the dataset in the datastorr registry. Defaults to NULL.}
}
\description{
This function deletes a dataset release from GitHub, using the datastorr package.
}
\examples{
dataset_del("v1.0")

}
