% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access.R
\name{dataset_versions}
\alias{dataset_versions}
\alias{dataset_version_current}
\title{Get Available Versions of the Australian Plant Census and Australian Plant Names Index Datasets}
\usage{
dataset_versions(local = TRUE, path = NULL)

dataset_version_current(local = TRUE, path = NULL)
}
\arguments{
\item{local}{Logical indicating if local or Github versions should be polled.
The default is TRUE, which means that only local versions will be checked.
If set to FALSE, the function will check for the most recent Github version
if there are no local versions available. Note that local versions take precedence
over Github versions if there are conflicts.}

\item{path}{Path to store the data at.}
}
\description{
This function returns a list of available versions of the Australian Plant Census
and Australian Plant Names Index datasets. The versions can be either local or
from the Github repository.
}
\examples{
# Get available local versions of the Australian Plant Census and Australian Plant Names Index datasets
dataset_versions()

# Get available Github versions of the Australian Plant Census and Australian Plant Names Index datasets
dataset_versions(local = FALSE)

# Get available versions of the Australian Plant Census and Australian Plant Names Index datasets stored at a specific path
dataset_versions(path = "/path/to/dataset")
}
