% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{download_taxonomic_resources_for_release}
\alias{download_taxonomic_resources_for_release}
\title{Download taxonomic resources for GitHub Release}
\usage{
download_taxonomic_resources_for_release(version_name = NULL, path = "ignore/")
}
\arguments{
\item{version_name}{character string of version name, follow semantic versioning}

\item{path}{to download parquets to upload}
}
\description{
Download taxonomic resources for GitHub Release
}
\keyword{internal}
