% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_names.R
\name{extract_genus}
\alias{extract_genus}
\title{Extract Genus}
\usage{
extract_genus(taxon_name)
}
\arguments{
\item{taxon_name}{}
}
\value{
The genus for a scientific name.
}
\description{
This function extracts the genus component of a scientific name.
It identifies if the genus is/is not a hybrid. For a hybrid genus,
the first two words of the taxon name are extracted (e.g. "x Cynochloris"),
while for a non-hybrid genus just the first word is extracted (e.g. "Banksia").
}
\examples{
genus = extract_genus(stripped_name)


}
