% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match taxonomic names}
\usage{
fuzzy_match(
  txt,
  accepted_list,
  max_distance_abs,
  max_distance_rel,
  n_allowed = 1
)
}
\arguments{
\item{txt}{The string of text requiring a match}

\item{accepted_list}{The list of accepted names attempting to match to}

\item{max_distance_abs}{The maximum allowable number of characters differing between the input string and the match}

\item{max_distance_rel}{The maximum proportional difference between the input string and the match}

\item{n_allowed}{The number of allowable matches returned. Defaults to 1}
}
\value{
A text string that matches a recognised taxon name or scientific name
}
\description{
This function attempts to match input strings to a list of allowable taxonomic names.
It requires that the first letter (or digit) of each word is identical between the input and output strings to avoid mis-matches
}
\examples{
fuzzy_match("Baksia serrata", c("Banksia serrata", "Banksia integrifolia"), max_distance_abs = 1, max_distance_rel = 1)

}
