% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_names.R
\name{standardise_taxon_rank}
\alias{standardise_taxon_rank}
\title{Standardise taxon ranks from latin into english.}
\usage{
standardise_taxon_rank(taxon_rank)
}
\arguments{
\item{taxon_rank}{A character vector of taxon ranks that need to be standardised.}
}
\value{
A character vector of standardised taxon names.
}
\description{
The function takes a character vector of taxon ranks as input and
returns a character vector of taxon ranks using standardised english terms.
}
\examples{
standardise_taxon_rank(c("regnum", "kingdom", "classis", "class"))
}
