% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_diversity.R
\name{state_diversity_counts}
\alias{state_diversity_counts}
\title{Calculate state-level diversity for native, introduced, and more complicated species origins}
\usage{
state_diversity_counts(
  state = c("NSW", "NT", "Qld", "WA", "ChI", "NSW", "SA", "Vic", "Tas", "ACT", "NI",
    "LHI", "MI", "HI", "MDI", "CoI", "CSI", "AR", "CaI"),
  resources = load_taxonomic_resources(version = default_version())
)
}
\arguments{
\item{state}{A character string indicating the Australian state or territory to calculate the diversity for. Default is "NSW". Possible values are "NSW", "NT", "Qld", "WA", "ChI", "SA", "Vic", "Tas", "ACT", "NI", "LHI", "MI", "HI", "MDI", "CoI", "CSI", and "AR".}

\item{resources}{the taxonomic resources to make the summary statistics. this defaults to default_version().  call load_taxonomic_resources seperately to greatly speed this function up.}
}
\value{
A tibble of diversity counts for the specified state or territory, including native, introduced, and more complicated species origins.
The tibble has three columns: "origin" indicating the origin of the species, "state" indicating the Australian state or territory, and "num_species" indicating the number of species for that origin and state.
}
\description{
This function calculates state-level diversity for native, introduced, and more complicated species origins based on the geographic data available in the current Australian Plant Census.
}
\examples{
 \dontrun{state_diversity_counts(state = "NSW")}
}
