% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms_for_accepted_names.R
\name{synonyms_for_accepted_names}
\alias{synonyms_for_accepted_names}
\title{Synonyms for Currently Accepted Names}
\usage{
synonyms_for_accepted_names(
  resources = load_taxonomic_resources(),
  accepted_names,
  collapse = TRUE
)
}
\arguments{
\item{resources}{}

\item{collapse}{Offering the option to return a long data table with each synonym in its own row,
versus collapsed into a vector for each accepted name}
}
\value{
A table with the currently accepted name and columns documenting all synonyms and all synonyms with taxonomic status.
}
\description{
This function generates lists a string of synonyms for currently accepted names to facilitate working out past names of a taxon
when the current name is known
}
\examples{
synonyms_for_accepted_names(resources = resources, accepted_names = c("Justicia tenella", "Acacia aneura"), collapse = TRUE)
synonyms_for_accepted_names()

}
