% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{word}
\alias{word}
\title{Extract words from a sentence. Intended as a faster
replacement for stringr::word}
\usage{
word(string, start = 1L, end = start, sep = " ")
}
\arguments{
\item{string}{A character vector}

\item{start, end}{Pair of integer vectors giving range of words (inclusive)
to extract. The default value select the first word.}

\item{sep}{Separator between words. Defaults to single space.}
}
\value{
A character vector with the same length as \code{string}/\code{start}/\code{end}.
}
\description{
Extract words from a sentence. Intended as a faster
replacement for stringr::word
}
\examples{
spp <- c("Banksia serrata", "Actinotus helanthii")
APCalign:::word(spp, 1)
APCalign:::word(spp, 2)
}
