% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{align_taxa}
\alias{align_taxa}
\title{Find taxonomic alignments for a list of names}
\usage{
align_taxa(
  original_name,
  output = NULL,
  fuzzy_matching = TRUE,
  max_distance_abs = 3,
  max_distance_rel = 0.2,
  ver = default_version()
)
}
\arguments{
\item{original_name}{list of names to query}

\item{output}{(optional) name of file to save results}

\item{fuzzy_matching}{option to turn off fuzzy matching}

\item{max_distance_abs}{absolute distance in substitution space}

\item{max_distance_rel}{relative distance in substitution space}

\item{ver}{version number for flora resource}
}
\description{
Use APC & APNI to find taxonomic alignments for a list of names
}
