% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{align_taxa}
\alias{align_taxa}
\title{Find taxonomic alignments for a list of names}
\usage{
align_taxa(
  original_name,
  output = NULL,
  fuzzy_matching = FALSE,
  max_distance_abs = 3,
  max_distance_rel = 0.2,
  resources = load_taxonomic_resources()
)
}
\arguments{
\item{original_name}{A list of names to query for taxonomic alignments.}

\item{output}{(optional) The name of the file to save the results to.}

\item{fuzzy_matching}{An option to turn off fuzzy matching.}

\item{max_distance_abs}{The absolute distance in substitution space.}

\item{max_distance_rel}{The relative distance in substitution space.}

\item{resources}{XXXX}
}
\value{
A tibble with columns: original_name, cleaned_name, aligned_name, source, known, and checked.
}
\description{
This function uses APC & APNI to find taxonomic alignments for a list of names.
}
\examples{
align_taxa(c("Poa annua", "Abies alba"), output = "taxa_alignments.csv")

}
\seealso{
\code{\link{default_version}} for the default value of the \code{ver} parameter.
}
\concept{taxonomic alignment functions}
\keyword{APC,}
\keyword{APNI,}
\keyword{alignments,}
\keyword{flora}
\keyword{resource}
\keyword{taxonomic}
