% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_taxa.R
\name{align_taxa}
\alias{align_taxa}
\title{Find taxonomic alignments for a list of names to a version of the Australian Plant Census (APC) through standardizing formatting and checking for spelling issues}
\usage{
align_taxa(
  original_name,
  output = NULL,
  resources = load_taxonomic_resources(),
  fuzzy_abs_dist = 3,
  fuzzy_rel_dist = 0.2,
  fuzzy_matches = TRUE,
  imprecise_fuzzy_matches = FALSE,
  APNI_matches = TRUE,
  identifier = NA_character_
)
}
\arguments{
\item{original_name}{A list of names to query for taxonomic alignments.}

\item{output}{(optional) The name of the file to save the results to.}

\item{resources}{the taxonomic resources used to align the taxa names. Loading this can be slow,
so call \code{\link{load_taxonomic_resources}} separately to greatly speed this function up and pass the resources in.}

\item{fuzzy_abs_dist}{The number of characters allowed to be different for a fuzzy match.}

\item{fuzzy_rel_dist}{The proportion of characters allowed to be different for a fuzzy match.}

\item{fuzzy_matches}{Fuzzy matches are turned on as a default. The relative and absolute distances allowed for fuzzy matches to species and infraspecific taxon names are defined by the parameters \code{fuzzy_abs_dist} and \code{fuzzy_rel_dist}}

\item{imprecise_fuzzy_matches}{Imprecise fuzzy matches are turned off as a default.}

\item{APNI_matches}{Name matches to the APNI (Australian Plant Names Index) are turned off as a default.}

\item{identifier}{A dataset, location or other identifier, which defaults to NA.}
}
\value{
A tibble with columns: original_name, cleaned_name, aligned_name, source, known, and checked.
}
\description{
This function uses Australian Plant Census (APC) & the Australian Plant Name Index (APNI) to find taxonomic alignments for a list of names.
}
\examples{
\donttest{align_taxa(c("Poa annua", "Abies alba"))}

}
\seealso{
\code{\link{load_taxonomic_resources}}

Other taxonomic alignment functions: 
\code{\link{create_taxonomic_update_lookup}()},
\code{\link{update_taxonomy}()}
}
\concept{taxonomic alignment functions}
