% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_taxa.R
\name{align_taxa}
\alias{align_taxa}
\title{Find taxonomic alignments for a list of names to a version of the Australian Plant Census (APC) through standardizing formatting and checking for spelling issues}
\usage{
align_taxa(
  original_name,
  output = NULL,
  max_distance_abs = 3,
  max_distance_rel = 0.2,
  resources = load_taxonomic_resources()
)
}
\arguments{
\item{original_name}{A list of names to query for taxonomic alignments.}

\item{output}{(optional) The name of the file to save the results to.}

\item{max_distance_abs}{The absolute distance in substitution space for the fuzzy matching.}

\item{max_distance_rel}{The relative distance in substitution space for the fuzzy matching.}

\item{resources}{the taxonomic resources used to align the taxa names. Loading this can be slow,
so call \code{\link{load_taxonomic_resources}} separately to greatly speed this function up and pass the resources in.}
}
\value{
A tibble with columns: original_name, cleaned_name, aligned_name, source, known, and checked.
}
\description{
This function uses Australian Plant Census (APC) & the Australian Plant Name Index (APNI) to find taxonomic alignments for a list of names.
}
\examples{
align_taxa(c("Poa annua", "Abies alba"))

}
\seealso{
\code{\link{load_taxonomic_resources}}

Other taxonomic alignment functions: 
\code{\link{create_taxonomic_update_lookup}()},
\code{\link{update_taxonomy}()}
}
\concept{taxonomic alignment functions}
