% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxonomic_resources.R
\name{default_version}
\alias{default_version}
\title{Get the default version for stable data}
\usage{
default_version()
}
\value{
A character string representing the default version for stable data.
}
\description{
This function returns the default version for stable data, which is used when no
version is specified.
}
\examples{
default_version()

}
