% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_diversity.R
\name{get_apc_genus_family_lookup}
\alias{get_apc_genus_family_lookup}
\title{Genus-Family Lookup}
\usage{
get_apc_genus_family_lookup(type_of_data = "stable", ver = default_version())
}
\description{
This function returns a genus family lookup from a stable or live version of the APC.
}
