% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxonomic_resources.R
\name{get_versions}
\alias{get_versions}
\title{Which versions of taxonomic resources are available?}
\usage{
get_versions()
}
\value{
tibble of dates when APC/APNI resources were downloaded as a Github Release
}
\description{
Which versions of taxonomic resources are available?
}
\examples{
get_versions()
}
