% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{load_taxonomic_resources}
\alias{load_taxonomic_resources}
\title{Title}
\usage{
load_taxonomic_resources(ver = default_version(), reload = FALSE)
}
\arguments{
\item{reload}{}
}
\value{

}
\description{
Title
}
