% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{load_taxonomic_resources}
\alias{load_taxonomic_resources}
\title{Load taxonomic resources}
\usage{
load_taxonomic_resources(
  version = default_version(),
  reload = FALSE,
  filetype = "parquet"
)
}
\arguments{
\item{ver}{The version number of the dataset to use. Defaults to the default version.}

\item{reload}{A logical indicating whether to reload the dataset from the data source. Defaults to FALSE.}

\item{filetype}{type of file to download. parquet or csv}
}
\value{
The taxonomic resources data loaded into the global environment.
}
\description{
Loads taxonomic resources into the global environment. This function accesses taxonomic data from a dataset using the provided version number or the default version. The loaded data contains two lists: APC and APNI, which contain taxonomic information about plant species in Australia. The function creates several data frames by filtering and selecting data from the loaded lists.
}
\examples{
load_taxonomic_resources()


}
