% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_diversity.R
\name{native_anywhere_in_australia}
\alias{native_anywhere_in_australia}
\title{Check If Species Is Native Anywhere In Australia}
\usage{
native_anywhere_in_australia(species, resources = load_taxonomic_resources())
}
\arguments{
\item{species}{A character string representing the binomial or common name of the species.}

\item{resources}{An optional list of taxonomic resources to use for the lookup.
If not provided, the function will load default taxonomic resources using the `load_taxonomic_resources()` function.}
}
\value{
A tibble with two columns: `species`, which is the same as the input `species`,
        and `native_anywhere_in_aus`, a logical vector indicating whether each species is native anywhere in Australia.
}
\description{
This function checks if the given species is native anywhere in Australia.
It creates a lookup table from taxonomic resources, and checks if the species
is listed as native in that table.
}
\examples{
resources<-load_taxonomic_resources()
native_anywhere_in_australia("Eucalyptus globulus")
native_anywhere_in_australia(c("Eucalyptus globulus","Pinus radiata"), resources=resources)
}
