% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{strip_names}
\alias{strip_names}
\title{Title}
\usage{
strip_names(taxon_names)
}
\arguments{
\item{taxon_names}{a vector of names to strip}
}
\value{
a vector of stripped names
}
\description{
Title
}
