% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Title}
\usage{
update_taxonomy(
  aligned_names,
  output = "taxon_list.csv",
  ver = default_version()
)
}
\arguments{
\item{ver}{}
}
\value{

}
\description{
Title
}
