% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_taxonomy.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Use APC and APNI to update taxonomy, replacing synonyms to current taxa where relevant}
\usage{
update_taxonomy(
  aligned_data,
  taxonomic_splits = "most_likely_species",
  output = NULL,
  resources = load_taxonomic_resources()
)
}
\arguments{
\item{aligned_data}{A tibble of plant names to update. This table must include 5 columns, original_name, aligned_name, taxon_rank, taxonomic_dataset, and aligned_reason.
These columns are created by the function \code{align_taxa}.
The columns \code{original_name} and \code{aligned_name} must be in the format of the scientific name, with genus and species,
and may contain additional qualifiers such as subspecies or varieties. The names are case insensitive.}

\item{taxonomic_splits}{XXXXX}

\item{output}{(optional) Name of the file where results are saved. The default is NULL and no file is created.
If specified, the output will be saved in a CSV file with the given name.}

\item{resources}{the taxonomic resources required to make the summary statistics.  Loading this can be slow, so call load_taxonomic_resources separately to greatly speed this function up and pass the resources in.}
}
\value{
A tibble with updated taxonomy for the specified plant names. The tibble contains the following columns:
\itemize{
\item \code{original_name}: the original plant name.
\item \code{aligned_name}: the input plant name.
\item \code{accepted_name}: the APC-accepted plant name, when available.
\item \code{suggested_name}: the suggested plant name to use. Identical to the accepted_name, when an accepted_name exists.
\item \code{taxonomic_dataset}: the source of the updated taxonomic information (APC or APNI).
\item \code{accepted_name_usage_ID}: the unique identifier for the accepted name of the input name.
\item \code{canonical_name}: the accepted (or known) scientific name for the input name.
\item \code{scientific_name_authorship}: the authorship information for the accepted (or known) name.
\item \code{taxon_rank}: the taxonomic rank of the accepted (or known) name.
\item \code{taxonomic_status}: the taxonomic status of the accepted (or known) name.
\item \code{ taxonomic_status_aligned}: for taxa where there is ambiguity due to a taxon split, the taxonomic status of each possible match.
\item \code{aligned_reason}: the explanation of a specific taxon name alignment (from an original name to an aligned name).
\item \code{update_reason}: the explanation of a specific taxon name update (from an aligned name to an accepted name).
\item \code{genus}: the genus of the accepted (or known) name.
\item \code{family}: the family of the accepted (or known) name.
\item \code{subclass}: the subclass of the accepted name.
\item \code{taxon_distribution}: the distribution of the accepted name.
\item \code{taxon_ID}: an identifier for a specific taxon concept.
\item \code{taxon_ID_genus}: an identifier for the genus.
\item \code{scientific_name_ID}: an identifier for the nomenclatural (not taxonomic) details of a scientific name.
\item \code{row_number}: the row number of a specific original_name in the input.
}
}
\description{
This function uses the Australia's Virtual Herbarium's taxonomic resources, specifically the Australian Plant
Census (APC) and the Australian Plant Name Index (APNI), to update taxonomy of plant species, replacing any synonyms
to their current accepted name.
}
\examples{
# Update taxonomy for two plant names and print the result
update_taxonomy(
 tibble::tibble(
   original_name = c("Dryandra preissii", "Banksia acuminata"),
   aligned_name = c("Dryandra preissii", "Banksia acuminata"),
   taxon_rank = c("Species", "Species"),
   taxonomic_dataset = c("APC", "APC"),
   aligned_reason = NA_character_
 )
)
}
\seealso{
load_taxonomic_resources

Other taxonomic alignment functions: 
\code{\link{align_taxa}()},
\code{\link{create_taxonomic_update_lookup}()}
}
\concept{taxonomic alignment functions}
