% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_names.R
\name{update_taxonomy}
\alias{update_taxonomy}
\title{Use APC & APNI to update taxonomy, replacing synonyms to current taxa where relevant}
\usage{
update_taxonomy(
  aligned_names,
  output = "taxon_list.csv",
  ver = default_version()
)
}
\arguments{
\item{output}{(optional) Name of file where results are saved}

\item{ver}{}
}
\value{

}
\description{
Use APC & APNI to update taxonomy, replacing synonyms to current taxa where relevant
}
